Rename SenderBubble and OtherUserBubble
These are no longer the actual bubbles but wrappers for various message layouts depending on number of attachments.
