+++ client/src/components/ActiveChat/Messages.js
@@ -1,18 +1,32 @@
 import React from "react";
 import { Box } from "@material-ui/core";
-import { SenderBubble, OtherUserBubble } from "../ActiveChat";
+import SenderMessage from "./SenderMessage";
+import OtherUserMessage from "./OtherUserMessage";
+
 import moment from "moment";
 
 const Messages = (props) => {
   const { messages, otherUser, userId } = props;
+  
+  console.log(`Messages component has data: `, messages);
+  
   return (
     <Box>
       {messages.map((message) => {
         const time = moment(message.createdAt).format("h:mm");
         return message.senderId === userId ? (
-          <SenderBubble key={message.id} text={message.text} time={time} />
+          <SenderMessage 
+            key={message.id} 
+            message={message} 
+            time={time} 
+          />
         ) : (
-          <OtherUserBubble key={message.id} text={message.text} time={time} otherUser={otherUser} />
+          <OtherUserMessage 
+            key={message.id} 
+            message={message} 
+            time={time} 
+            otherUser={otherUser} 
+          />
         );
       })}
     </Box>
